+++ nodeapp/src/__tests__/integration/v2/v1/news.route.test.ts
@@ -2,8 +2,8 @@ import { Express } from 'express';
 import request from "supertest";
 import { newDb, IMemoryDb } from 'pg-mem';
 import { DataSource } from 'typeorm';
-import { App } from '../../../../app';
-import { News } from '../../../../models/news.entity';
+import { App } from '../../../app';
+import { News } from '../../../models/news.entity';
 
 
 let app: App;
