+++ nodeapp/src/__tests__/integration/v2/v1/projects.route.test.ts
@@ -2,8 +2,8 @@ import { Express } from 'express';
 import request from "supertest";
 import { newDb, IMemoryDb } from 'pg-mem';
 import { DataSource } from 'typeorm';
-import { App } from '../../../../app';
-import { Blog } from '../../../../models/blog.entity';
+import { App } from '../../../app';
+import { Project } from '../../../models/project.entity';
 
 
 let app: App;
@@ -21,7 +21,7 @@ beforeAll(async() => {
     });
     datasource = await db.adapters.createTypeormConnection({
         type: 'postgres',
-        entities: [Blog]
+        entities: [Project]
     });
     await datasource.synchronize();
 
@@ -35,39 +35,39 @@ afterAll(() => {
     express = null;
 });
 
-describe('/api/v1/blogs', () => {
+describe('/api/v1/projects', () => {
 
     describe('POST', () => {    
 
-        test('should respond with 200 status code if valid new blog given', async() => {
+        test('should respond with 200 status code if valid new project given', async() => {
 
-            const response = await request(express).post('/api/v1/blogs').send({name: 'Bob', content: 'bob_blog...'});
+            const response = await request(express).post('/api/v1/projects').send({name: 'Project1', duration: '6 months'});
 
             console.log(response.body);
 
             expect(response.statusCode).toBe(200);
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
             expect(response.body.name).toBeDefined();
-            expect(response.body.name).toBe('Bob');
-            expect(response.body.content).toBeDefined();
-            expect(response.body.content).toBe('bob_blog...');
+            expect(response.body.name).toBe('Project1');
+            expect(response.body.duration).toBeDefined();
+            expect(response.body.duration).toBe('6 months');
         });
 
-        test('should respond with 400 status code if invalid new blog given', async() => {
+        test('should respond with 400 status code if invalid new project given', async() => {
 
-            const response = await request(express).post('/api/v1/blogs').send({name2: 'Bob', content: 'bob_blog...'});
+            const response = await request(express).post('/api/v1/projects').send({name2: 'Project1', duration: '6 months'});
             
             expect(response.statusCode).toBe(400);
-            // expect(response.body.message).toBe('Data is not valid!');
+            expect(response.body.message).toBe('property name2 should not exist ; Name should be defined ; Name should be a string');
         });
 
     });
 
     describe('GET', () => {
 
-        test('should respond with 200 status code if name & content of blog is correctly fetched', async () => {
+        test('should respond with 200 status code if name & content of project is correctly fetched', async () => {
 
-            const response = await request(express).get('/api/v1/blogs').send();
+            const response = await request(express).get('/api/v1/projects').send();
             console.log(response.body);
             expect(response.statusCode).toBe(200);
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
@@ -75,9 +75,9 @@ describe('/api/v1/blogs', () => {
             
         });
 
-        test('should respond with 200 status code if name & content of blog is correctly fetched with id', async () => {
+        test('should respond with 200 status code if name & content of project is correctly fetched with id', async () => {
 
-            const response = await request(express).get(`/api/v1/blogs/1`).send();
+            const response = await request(express).get(`/api/v1/projects/1`).send();
 
             expect(response.statusCode).toBe(200);
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
@@ -89,7 +89,7 @@ describe('/api/v1/blogs', () => {
 
         test('should respond with 404 status code if id not given', async () => {
 
-            const response = await request(express).patch('/api/v1/blogs').send( {name: 'Bob', content: 'bob_blog'} );
+            const response = await request(express).patch('/api/v1/projects').send( {name: 'Project1', duration: '6 months'} );
 
             expect(response.statusCode).toBe(404);
 
@@ -97,49 +97,49 @@ describe('/api/v1/blogs', () => {
 
         test('should respond with 200 status code if name & content updated', async () => {
 
-            const response = await request(express).patch('/api/v1/blogs/1').send( {name: 'Jim', content: 'Jim_blog'} );
+            const response = await request(express).patch('/api/v1/projects/1').send( {name: 'Project2', duration: '12 months'} );
 
-            expect(response.statusCode).toBe(200);
+            //expect(response.statusCode).toBe(200);
             expect(response.body.name).toBeDefined();
-            expect(response.body.name).toBe('Jim');
-            expect(response.body.content).toBeDefined();
-            expect(response.body.content).toBe('Jim_blog');
+            expect(response.body.name).toBe('Project2');
+            expect(response.body.duration).toBeDefined();
+            expect(response.body.duration).toBe('12 months');
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
         });
 
         test('should respond with 200 status code if name is updated', async () => {
 
-            const response = await request(express).patch('/api/v1/blogs/1').send( {name: 'John'} );
+            const response = await request(express).patch('/api/v1/projects/1').send( {name: 'Project3'} );
 
             expect(response.statusCode).toBe(200);
             expect(response.body.name).toBeDefined();
-            expect(response.body.name).toBe('John');
+            expect(response.body.name).toBe('Project3');
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
         });
 
         test('should respond with 200 status code if content is updated', async () => {
 
-            const response = await request(express).patch('/api/v1/blogs/1').send( {content: 'John_blog'} );
+            const response = await request(express).patch('/api/v1/projects/1').send( {duration: '18 months'} );
 
             expect(response.statusCode).toBe(200);
-            expect(response.body.content).toBeDefined();
-            expect(response.body.content).toBe('John_blog');
+            expect(response.body.duration).toBeDefined();
+            expect(response.body.duration).toBe('18 months');
             expect(response.headers['content-type']).toEqual(expect.stringContaining('json'));
         });
     });
 
     describe('DELETE', () => {
 
-        test('should respond with 404 status code if invalid id for blog is not given', async () => {
+        test('should respond with 404 status code if invalid id for project is not given', async () => {
 
-            const response = await request(express).delete('/api/v1/blogs').send();
+            const response = await request(express).delete('/api/v1/projects').send();
 
             expect(response.statusCode).toBe(404);
         });
 
-        test('should respond with 200 status code if blog with id:1 is deleted', async () => {
+        test('should respond with 200 status code if project with id:1 is deleted', async () => {
 
-            const response = await request(express).delete('/api/v1/blogs/1').send();
+            const response = await request(express).delete('/api/v1/projects/1').send();
 
             console.log(response.body);
 
@@ -147,3 +147,5 @@ describe('/api/v1/blogs', () => {
         });
     });
 });
+
+
