+++ nodeapp/src/__tests__/unit/services/blogs.service.test.ts
@@ -1,0 +1,32 @@
+import { Blog } from '../../../models/blog.entity';
+import { Service } from '../../../services/index.service';
+
+
+beforeAll(async() => {
+
+    // const blogService = new Service(Blog, DataSource);
+});
+
+
+describe('blogs service', () => {
+
+    test('create a blog', async() => {
+
+    });
+
+    test('find all blogs', async() => {
+
+    });
+
+    test('find a blog', async() => {
+
+    });
+
+    test('patch a blog', async() => {
+
+    });
+
+    test('delete a blog', async() => {
+
+    });
+});
