+++ nodeapp/src/app.ts
@@ -54,15 +54,25 @@ export class App {
     }
     
     private registerRoutes( ) {
-        this.express.use( `/api/v1/blogs`, new Route(BlogValidator, new Service(this.datasource.getRepository(Blog))).router);
-        this.express.use( `/api/v1/projects`, new Route(ProjectValidator, new Service(this.datasource.getRepository(Project))).router);
-        this.express.use( `/api/v1/weather`, new Route(WeatherValidator, new WeatherService(this.datasource.getRepository(Weather))).router); 
-        this.express.use( `/api/v1/news`, new Route(NewsValidator, new NewsService(this.datasource.getRepository(News))).router);
-        this.express.use( `/api/v1/books`, new Route(BookValidator, new Service(this.datasource.getRepository(Book))).router);
-        this.express.use( `/api/v1/products`, new Route(ProductValidator, new Service(this.datasource.getRepository(Product))).router);
-        this.express.use( `/api/v1/users`, new Route(UserValidator, new Service(this.datasource.getRepository(User))).router);
-        this.express.use( `/api/v1/auth`, new AuthV1Route(AuthUserValidator, new Service(this.datasource.getRepository(User))).router);
-        this.express.use( `/api/v2/auth`, new AuthV2Route(AuthUserValidator, new Service(this.datasource.getRepository(User))).router);
+        this.express.use( `/api/v1/blogs`, new Route(BlogValidator, new Service(Blog, this.datasource)).router);
+        this.express.use( `/api/v1/projects`, new Route(ProjectValidator, new Service(Project, this.datasource)).router);
+        this.express.use( `/api/v1/weather`, new Route(WeatherValidator, new WeatherService(Weather, this.datasource)).router); 
+        this.express.use( `/api/v1/news`, new Route(NewsValidator, new NewsService(News, this.datasource)).router);
+        this.express.use( `/api/v1/books`, new Route(BookValidator, new Service(Book, this.datasource)).router);
+        this.express.use( `/api/v1/products`, new Route(ProductValidator, new Service(Product, this.datasource)).router);
+        this.express.use( `/api/v1/users`, new Route(UserValidator, new Service(User, this.datasource)).router);
+        this.express.use( `/api/v1/auth`, new AuthV1Route(AuthUserValidator, new Service(User, this.datasource)).router);
+        this.express.use( `/api/v2/auth`, new AuthV2Route(AuthUserValidator, new Service(User, this.datasource)).router);
+
+        // this.express.use( `/api/v1/blogs`, new Route(BlogValidator, new Service(this.datasource.getRepository(Blog))).router);
+        // this.express.use( `/api/v1/projects`, new Route(ProjectValidator, new Service(this.datasource.getRepository(Project))).router);
+        // this.express.use( `/api/v1/weather`, new Route(WeatherValidator, new WeatherService(this.datasource.getRepository(Weather))).router); 
+        // this.express.use( `/api/v1/news`, new Route(NewsValidator, new NewsService(this.datasource.getRepository(News))).router);
+        // this.express.use( `/api/v1/books`, new Route(BookValidator, new Service(this.datasource.getRepository(Book))).router);
+        // this.express.use( `/api/v1/products`, new Route(ProductValidator, new Service(this.datasource.getRepository(Product))).router);
+        // this.express.use( `/api/v1/users`, new Route(UserValidator, new Service(this.datasource.getRepository(User))).router);
+        // this.express.use( `/api/v1/auth`, new AuthV1Route(AuthUserValidator, new Service(this.datasource.getRepository(User))).router);
+        // this.express.use( `/api/v2/auth`, new AuthV2Route(AuthUserValidator, new Service(this.datasource.getRepository(User))).router);
     }
     
     private registerErrorHandlers( ) {
